+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
-## New in 0.4.0 (Released ??)
-* Show cumulative averages
+## New in 0.4.0 (Released 2017/11/28)
+* Show cumulative averages.
+* Reduce frequency of wickets.
 
 ## New in 0.3.0 (Released 2017/11/05)
 * Make each innings collapsable.